+++ CHANGELOG.md
@@ -9,3 +9,4 @@ HEAD
 * Add first MQTTClient class and include ArduinoJson lib dep
 * Add release github action and build scripts
 * Support for esp8266 target using PubSubClient, WebSocketStreamClient and ArduinoHttpClient for WebSockets client
+* Fix PubSubclient callback and include onDataReceived